Separate the initscale zero test
